Ignore endpoint events if processed already
In deployments, if we delete submariner-gateway machineset and then
later delete the submariner namespace in a cluster. It would lead
to a stale endpoint in other clusters. Now if we create a machineset
and join the cluster again, the old endpoint gets cleared. But the old
endpoint delete event comes after the new endpoint is created and the
host network routing rules gets deleted.

With this fix, a delete event will be ignored if it has timestamp older
than the current endpoint.

Fixes:1544

Signed-off-by: Aswin Surayanarayanan <asuryana@redhat.com>